requirements.txt
requests
flask
flask_restx
Pillow